LureScript and Physical torch
the lure script copies what you had for the player and is for the bait instead. So anything need to get the player attack to work will need to be applied likewise to the bait. The torch is shape objects in a empty parent object with a light and script to turn it on and off.