Item Interaction
Added a destroy function to the interactable base. Unsure how it works on other scripts, but it will run the OnInteract() function now and then check if there object has DestroyObj on it. If it does, it will gameObject destroy what the player is looking at once it is done executing OnInteract()